requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiomcache==0.8.2
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
async-lru==2.0.4
attrs==24.3.0
black==24.10.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
fastapi==0.115.6
flake8==7.1.1
frozenlist==1.5.0
h11==0.14.0
httptools==0.6.4
idna==3.10
isort==5.13.2
mccabe==0.7.0
multidict==6.1.0
mypy==1.14.1
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
propcache==0.2.1
pycodestyle==2.12.1
pydantic==2.10.5
pydantic_core==2.27.2
pydash==8.0.4
pyflakes==3.2.0
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
sniffio==1.3.1
starlette==0.41.3
tenacity==9.0.0
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
websockets==14.1
yarl==1.18.3
